new report: week for users (#2494)
