+++ src/Reporting/AbstractUserList.php
@@ -1,0 +1,28 @@
+<?php
+
+/*
+ * This file is part of the Kimai time-tracking app.
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace App\Reporting;
+
+abstract class AbstractUserList
+{
+    /**
+     * @var \DateTime
+     */
+    private $date;
+
+    public function getDate(): ?\DateTime
+    {
+        return $this->date;
+    }
+
+    public function setDate(\DateTime $date): void
+    {
+        $this->date = $date;
+    }
+}
