+++ src/Reporting/Report.php
@@ -14,12 +14,16 @@ final class Report implements ReportInterface
     private $id;
     private $label;
     private $route;
+    private $reportIcon = 'reporting';
 
-    public function __construct(string $id, string $route, string $label)
+    public function __construct(string $id, string $route, string $label, ?string $reportIcon = null)
     {
         $this->id = $id;
         $this->route = $route;
         $this->label = $label;
+        if (null !== $reportIcon) {
+            $this->reportIcon = $reportIcon;
+        }
     }
 
     public function getRoute(): string
@@ -35,5 +39,10 @@ final class Report implements ReportInterface
     public function getLabel(): string
     {
         return $this->label;
+    }
+
+    public function getReportIcon(): string
+    {
+        return $this->reportIcon;
     }
 }
