+++ templates/macros/search.html.twig
@@ -51,7 +51,7 @@
             </button>
             <ul class="dropdown-menu">
                 <li>
-                    <a href="?removeDefaultQuery={{ form.vars.data.bookmark.name }}" id="removeDefaultQuery">{{ 'action.delete'|trans }}</a>
+                    <a href="?removeDefaultQuery={{ form.vars.data.bookmark.name }}" id="removeDefaultQuery">{{ 'label.remove_default'|trans }}</a>
                 </li>
             </ul>
         {% else %}
