+++ translations/messages.en.xlf
@@ -66,7 +66,11 @@
             </trans-unit>
             <trans-unit id="label.set_as_default">
                 <source>label.set_as_default</source>
-                <target>Save as default setting</target>
+                <target>Save setting as search favourite</target>
+            </trans-unit>
+            <trans-unit id="label.remove_default">
+                <source>label.remove_default</source>
+                <target>Delete search favourite</target>
             </trans-unit>
             <trans-unit id="label.asc">
                 <source>label.asc</source>
