+++ translations/reporting.de.xlf
@@ -10,6 +10,10 @@
                 <source>report_user_month</source>
                 <target>Monatsansicht für einen Benutzer</target>
             </trans-unit>
+            <trans-unit id="report_weekly_users">
+                <source>report_weekly_users</source>
+                <target>Wochenansicht für alle Benutzer</target>
+            </trans-unit>
             <trans-unit id="report_monthly_users">
                 <source>report_monthly_users</source>
                 <target>Monatsansicht für alle Benutzer</target>
